+++ test/src/main/resources/applicationContext.xml
@@ -9,7 +9,7 @@
 	
 	<!-- base-package 설정위치 포함 하위 패키지 모두 검색 처리
 		여러개 선언해도 되고, 구분자(빈칸, 콤마, 세미콜른 등) -->
-	<context:component-scan base-package="com.spring.biz" />
+	<context:component-scan base-package="com.todoc" />
 
 	<!-- 여러개 선언 방식
 	<context:component-scan base-package="com.spring.biz.user" />
@@ -39,7 +39,7 @@
 		<property name="mapperLocations" value="classpath:mappings/**/*Mapper.xml"></property>
 		
 		<!-- alias 설정 : 해당위치 클래스명 대소문자 구분없이 별칭 사용 -->
-		<property name="typeAliasesPackage" value="com.spring.biz" />
+		<property name="typeAliasesPackage" value="com.todoc" />
 		
 	</bean>
 	
@@ -48,7 +48,7 @@
 	</bean>
 	
 	<!-- Mapper 인터페이스 사용 방식 적용 -->
-	<mybatis-spring:scan base-package="com.spring.biz.**.dao"/>
+	<mybatis-spring:scan base-package="com.todoc.**.dao"/>
 	
 </beans>
