+++ test/src/main/webapp/WEB-INF/config/presentation-layer.xml
@@ -28,19 +28,19 @@
 	</bean>
 	
 	<!-- 예외 처리를 위한 설정 -->
-	<bean id="exceptionResolver" class="org.springframework.web.servlet.handler.SimpleMappingExceptionResolver">
-		<property name="defaultErrorView" value="/common/error.jsp" />
-		<property name="exceptionMappings"> 
-			<props>
-				<prop key="IllegalArgumentException">
-					common/illegalArgError
-				</prop>
-				<prop key="NullPointerException">
-					common/nullPointerError
-				</prop>
-			</props>
-		</property>
-	</bean>
+<!-- 	<bean id="exceptionResolver" class="org.springframework.web.servlet.handler.SimpleMappingExceptionResolver"> -->
+<!-- 		<property name="defaultErrorView" value="/common/error.jsp" /> -->
+<!-- 		<property name="exceptionMappings">  -->
+<!-- 			<props> -->
+<!-- 				<prop key="IllegalArgumentException"> -->
+<!-- 					common/illegalArgError -->
+<!-- 				</prop> -->
+<!-- 				<prop key="NullPointerException"> -->
+<!-- 					common/nullPointerError -->
+<!-- 				</prop> -->
+<!-- 			</props> -->
+<!-- 		</property> -->
+<!-- 	</bean> -->
 	<!-- 스프링 ViewResolver 등록 -->
 	<bean id="viewResolver"
 		class="org.springframework.web.servlet.view.InternalResourceViewResolver">
